persist login component used as main layout for outlet
