+++ client/src/components/Navbar.tsx
@@ -9,7 +9,6 @@ const Navbar = ({
     toggleAuthModal: MouseEventHandler<HTMLButtonElement>;
 }) => {
     const { auth } = useAuth();
-    console.log(auth);
 
     return (
         <header className="py-4 bg-primary text-white">
@@ -36,7 +35,7 @@ const Navbar = ({
                     </div>
                     {auth?.accessToken ? (
                         <Link
-                            to="/profile"
+                            to="/dashboard"
                             className="backdrop-blur-sm bg-white bg-opacity-40 text-white py-2 px-3 rounded-full font-semibold hover:shadow-white hover:[text-shadow:_0_2px_3px_rgb(0_0_0_/_40%)] transition-all"
                         >
                             <IconUser />
