+++ client/src/components/PersistLogin.tsx
@@ -3,13 +3,23 @@ import { useState, useEffect } from "react";
 import useRefreshToken from "../hooks/useRefreshToken";
 import useAuth from "../hooks/useAuth";
 import Loader from "./Loader";
+import Navbar from "./Navbar";
+import AuthModal from "./AuthModal";
 
 const PersistLogin = () => {
     const location = useLocation();
     const [isLoading, setIsLoading] = useState(true);
     const refresh = useRefreshToken();
     const { auth } = useAuth();
+    const [isAuthModal, setIsAuthModal] = useState(false);
 
+    const toggleAuthModal = () => {
+        setIsAuthModal((prev) => !prev);
+    };
+
+    /**
+     * Check if user is loged in
+     */
     useEffect(() => {
         const verifyRefreshToken = async () => {
             try {
@@ -26,7 +36,26 @@ const PersistLogin = () => {
             : verifyRefreshToken();
     }, [location.pathname]);
 
-    return <>{isLoading ? <Loader /> : <Outlet />}</>;
+    if (isLoading) return <Loader />;
+
+    return (
+        <>
+            <Navbar toggleAuthModal={toggleAuthModal} />
+            <Outlet />
+
+            {isAuthModal ? (
+                <>
+                    <AuthModal
+                        toggleAuthModal={toggleAuthModal}
+                        isAuthModal={isAuthModal}
+                    />
+                    <div className="w-screen h-screen bg-black backdrop-blur-sm bg-opacity-20 fixed top-0 left-0"></div>
+                </>
+            ) : (
+                ""
+            )}
+        </>
+    );
 };
 
 export default PersistLogin;
