+++ client/src/pages/Dashboard.tsx
@@ -1,0 +1,5 @@
+const Dashboard = () => {
+    return <div>dashboard</div>;
+};
+
+export default Dashboard;
